Make xcparse step configurable (#2)
These changes introduce a range of changes in order to allow for configuration of the xcparse commands. Users are now able to choose what extraction operations they want to do, what filtering of results they want, & how to divide the attachments/screenshots. It also allows for verbose logging & saving the results to the Bitrise Deploy directory automatically.